+++ src/main/java/com/example/waggle/global/security/JwtAuthenticationFilter.java
@@ -1,5 +1,6 @@
 package com.example.waggle.global.security;
 
+import com.example.waggle.global.security.service.TokenService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
