+++ src/main/java/com/example/waggle/global/security/oauth2/CustomOAuth2UserService.java
@@ -5,9 +5,9 @@ import com.example.waggle.domain.member.entity.Role;
 import com.example.waggle.domain.member.repository.MemberRepository;
 import com.example.waggle.global.exception.handler.AuthenticationHandler;
 import com.example.waggle.global.payload.code.ErrorStatus;
-import com.example.waggle.global.security.CustomUserDetails;
 import com.example.waggle.global.security.oauth2.OAuth2UserInfoFactory.AuthProvider;
 import com.example.waggle.global.security.oauth2.info.OAuth2UserInfo;
+import com.example.waggle.global.security.object.CustomUserDetails;
 import com.example.waggle.global.util.NameUtil;
 import com.example.waggle.global.util.NameUtil.NameType;
 import com.example.waggle.global.util.PasswordUtil;
