+++ src/main/java/com/example/waggle/global/security/object/CookieStatus.java
@@ -1,4 +1,4 @@
-package com.example.waggle.global.security;
+package com.example.waggle.global.security.object;
 
 public enum CookieStatus {
     EXIST_REFRESH, NOT_EXIST_REFRESH
