+++ src/main/java/com/example/waggle/global/util/NameUtil.java
@@ -42,5 +42,4 @@ public class NameUtil {
         usernameBuilder.append(prefix).append(uniqueValue);
         return usernameBuilder;
     }
-
 }
