+++ src/main/java/com/example/waggle/web/controller/TokenApiController.java
@@ -3,9 +3,9 @@ package com.example.waggle.web.controller;
 import com.example.waggle.global.exception.GeneralException;
 import com.example.waggle.global.payload.ApiResponseDto;
 import com.example.waggle.global.payload.code.ErrorStatus;
-import com.example.waggle.global.security.CookieStatus;
-import com.example.waggle.global.security.JwtToken;
-import com.example.waggle.global.security.TokenService;
+import com.example.waggle.global.security.object.CookieStatus;
+import com.example.waggle.global.security.object.JwtToken;
+import com.example.waggle.global.security.service.TokenService;
 import com.example.waggle.global.util.CookieUtil;
 import com.example.waggle.web.dto.member.MemberRequest;
 import io.swagger.v3.oas.annotations.Operation;
